+++ uamqp/mgmt_operation.py
@@ -8,6 +8,7 @@ import logging
 import uuid
 
 # from uamqp.session import Session
+from uamqp import Message
 from uamqp import constants
 from uamqp import errors
 from uamqp import c_uamqp
@@ -116,7 +117,7 @@ class MgmtOperation:
             if result != constants.MgmtExecuteResult.Ok:
                 _logger.error("Failed to complete mgmt operation.\nStatus code: {}\nMessage: {}".format(
                     status_code, description))
-            self._responses[operation_id] = wrapped_message
+            self._responses[operation_id] = Message(message=wrapped_message)
 
         self._mgmt_op.execute(operation, op_type, None, message.get_message(), on_complete)
         while not self._responses[operation_id] and not self.mgmt_error: